Module aws_sdk_mediapackagevod::model

Modules

asset_shallow

See AssetShallow

authorization

See Authorization

cmaf_encryption

See CmafEncryption

cmaf_package

See CmafPackage

dash_encryption

See DashEncryption

dash_manifest

See DashManifest

dash_package

See DashPackage

egress_access_logs

See EgressAccessLogs

egress_endpoint

See EgressEndpoint

hls_encryption

See HlsEncryption

hls_manifest

See HlsManifest

hls_package

See HlsPackage

mss_encryption

See MssEncryption

mss_manifest

See MssManifest

mss_package

See MssPackage

packaging_configuration

See PackagingConfiguration

packaging_group

See PackagingGroup

speke_key_provider

See SpekeKeyProvider

stream_selection

See StreamSelection

Structs

AssetShallow

A MediaPackage VOD Asset resource.

Authorization

CDN Authorization credentials

CmafEncryption

A CMAF encryption configuration.

CmafPackage

A CMAF packaging configuration.

DashEncryption

A Dynamic Adaptive Streaming over HTTP (DASH) encryption configuration.

DashManifest

A DASH manifest configuration.

DashPackage

A Dynamic Adaptive Streaming over HTTP (DASH) packaging configuration.

EgressAccessLogs

Configure egress access logging.

EgressEndpoint

The endpoint URL used to access an Asset using one PackagingConfiguration.

HlsEncryption

An HTTP Live Streaming (HLS) encryption configuration.

HlsManifest

An HTTP Live Streaming (HLS) manifest configuration.

HlsPackage

An HTTP Live Streaming (HLS) packaging configuration.

MssEncryption

A Microsoft Smooth Streaming (MSS) encryption configuration.

MssManifest

A Microsoft Smooth Streaming (MSS) manifest configuration.

MssPackage

A Microsoft Smooth Streaming (MSS) PackagingConfiguration.

PackagingConfiguration

A MediaPackage VOD PackagingConfiguration resource.

PackagingGroup

A MediaPackage VOD PackagingGroup resource.

SpekeKeyProvider

A configuration for accessing an external Secure Packager and Encoder Key Exchange (SPEKE) service that will provide encryption keys.

StreamSelection

A StreamSelection configuration.

Enums

AdMarkers

EncryptionMethod

ManifestLayout

PeriodTriggersElement

Profile

SegmentTemplateFormat

StreamOrder