Module aws_sdk_mediapackagevod::operation

Structs

ConfigureLogs

Changes the packaging group’s properities to configure log subscription

CreateAsset

Creates a new MediaPackage VOD Asset resource.

CreatePackagingConfiguration

Creates a new MediaPackage VOD PackagingConfiguration resource.

CreatePackagingGroup

Creates a new MediaPackage VOD PackagingGroup resource.

DeleteAsset

Deletes an existing MediaPackage VOD Asset resource.

DeletePackagingConfiguration

Deletes a MediaPackage VOD PackagingConfiguration resource.

DeletePackagingGroup

Deletes a MediaPackage VOD PackagingGroup resource.

DescribeAsset

Returns a description of a MediaPackage VOD Asset resource.

DescribePackagingConfiguration

Returns a description of a MediaPackage VOD PackagingConfiguration resource.

DescribePackagingGroup

Returns a description of a MediaPackage VOD PackagingGroup resource.

ListAssets

Returns a collection of MediaPackage VOD Asset resources.

ListPackagingConfigurations

Returns a collection of MediaPackage VOD PackagingConfiguration resources.

ListPackagingGroups

Returns a collection of MediaPackage VOD PackagingGroup resources.

ListTagsForResource

Returns a list of the tags assigned to the specified resource.

TagResource

Adds tags to the specified resource. You can specify one or more tags to add.

UntagResource

Removes tags from the specified resource. You can specify one or more tags to remove.

UpdatePackagingGroup

Updates a specific packaging group. You can’t change the id attribute or any other system-generated attributes.